packages/exo-observability/pyproject.toml
[project]
name = "exo-observability"
version = "0.1.0"
description = "Unified observability: structured logging, tracing, metrics, cost tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
]

[project.urls]
Homepage = "https://github.com/inclusionAI/AWorld"
Documentation = "https://inclusionai.github.io/AWorld/"
Repository = "https://github.com/inclusionAI/AWorld"
Issues = "https://github.com/inclusionAI/AWorld/issues"

[project.optional-dependencies]
otel = ["opentelemetry-api>=1.20", "opentelemetry-sdk>=1.20"]
otlp = ["opentelemetry-exporter-otlp>=1.20"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/exo"]
